Bringing SessionStatus and AppType to managed .NET
