+++ UtilitiesLibrary/Classes/Abstraction.cs
@@ -24,6 +24,9 @@ namespace WindowsUtils.Abstraction
                      .Select(f => f.GetValue(null))
                      .Cast<T>();
 
+        public static T GetById<T>(uint id) where T : Enumeration =>
+            GetAll<T>().First(f => f.Id == id);
+            
         public override bool Equals(object? obj)
         {
             if (obj is not Enumeration otherValue)
