+++ UtilitiesLibrary/Utilities.cs
@@ -11,29 +11,10 @@ using WindowsUtils.TerminalServices;
 #nullable enable
 namespace WindowsUtils
 {
-    public class SessionState : Enumeration
-    {
-        public static SessionState Active = new(0, "Active");
-        public static SessionState Connected = new(1, "Connected");
-        public static SessionState ConnectQuery = new(2, "ConnectQuery");
-        public static SessionState Shadow = new(3, "Shadow");
-        public static SessionState Disconnected = new(4, "Disconnected");
-        public static SessionState Idle = new(5, "Idle");
-        public static SessionState Listen = new(6, "Listen");
-        public static SessionState Reset = new(7, "Reset");
-        public static SessionState Down = new(8, "Down");
-        public static SessionState Init = new(9, "Init");
-        SessionState(uint id, string name) : base(id, name) { }
-
-        public static SessionState GetSessionStateById(uint id)
-        {
-            return GetAll<SessionState>().Where(q => q.Id == id).FirstOrDefault();
-        }
-    }
     public class Utilities
     {
         private static Session sessionInfo = new();
-
+        
         public static ComputerSession[] GetComputerSession(string computerName, bool onlyActive, bool IncludeSystemSession)
         {
             try { IPHostEntry hostEntry = Dns.GetHostEntry(computerName); }
